+++ src/components/StatisticsPanel/StatisticsDataRow.tsx
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-import { useState } from "react";
+import { useRef, useState } from "react";
 import Box from "@mui/material/Box";
 import IconButton from "@mui/material/IconButton";
 import ToggleButton from "@mui/material/ToggleButton";
@@ -36,9 +36,11 @@ import i18n from "@/i18n";
 import { makeStyles } from "@/util/styles";
 import { WithLocale } from "@/util/lang";
 import { isAreaStatistics, StatisticsRecord } from "@/model/statistics";
+import SnapshotButton from "@/components/SnapshotButton";
 import StatisticsTable from "./StatisticsTable";
 import HistogramChart from "./HistogramChart";
 import StatisticsRow from "./StatisticsRow";
+import { MessageType } from "@/states/messageLogState";
 
 const styles = makeStyles({
   table: {
@@ -53,6 +55,7 @@ interface StatisticsDataRowProps extends WithLocale {
   statisticsRecord: StatisticsRecord;
   rowIndex: number;
   removeStatistics: (rowIndex: number) => void;
+  postMessage: (messageType: MessageType, messageText: string | Error) => void;
 }
 
 export default function StatisticsDataRow({
@@ -60,7 +63,9 @@ export default function StatisticsDataRow({
   statisticsRecord,
   rowIndex,
   removeStatistics,
+  postMessage,
 }: StatisticsDataRowProps) {
+  const containerRef = useRef<HTMLDivElement>(null);
   const [brush, setBrush] = useState(false);
   const [details, setDetails] = useState(false);
   const { dataset, variable, time, placeInfo } = statisticsRecord.source;
@@ -80,6 +85,7 @@ export default function StatisticsDataRow({
       variable={variable}
       time={time}
       placeInfo={placeInfo}
+      containerRef={containerRef}
       actions={
         <>
           {hasHistogram && (
@@ -109,6 +115,14 @@ export default function StatisticsDataRow({
               </Tooltip>
             </ToggleButtonGroup>
           )}
+
+          {hasHistogram && (
+            <SnapshotButton
+              elementRef={containerRef}
+              postMessage={postMessage}
+            />
+          )}
+
           <IconButton size="small" onClick={handleRemoveStatistics}>
             <CloseIcon fontSize="inherit" />
           </IconButton>