+++ src/components/StatisticsPanel/StatisticsPanel.tsx
@@ -24,11 +24,12 @@
 
 import Box from "@mui/material/Box";
 
-import { makeStyles } from "@/util/styles";
+import { MessageType } from "@/states/messageLogState";
 import { StatisticsRecord } from "@/model/statistics";
 import { Dataset } from "@/model/dataset";
 import { Variable } from "@/model/variable";
 import { PlaceInfo } from "@/model/place";
+import { makeStyles } from "@/util/styles";
 import StatisticsDataRow from "./StatisticsDataRow";
 import StatisticsFirstRow from "./StatisticsFirstRow";
 
@@ -50,6 +51,7 @@ interface StatisticsPanelProps {
   statisticsRecords: StatisticsRecord[];
   addStatistics: () => void;
   removeStatistics: (index: number) => void;
+  postMessage: (messageType: MessageType, messageText: string | Error) => void;
 }
 
 export default function StatisticsPanel({
@@ -61,6 +63,7 @@ export default function StatisticsPanel({
   statisticsRecords,
   addStatistics,
   removeStatistics,
+  postMessage,
 }: StatisticsPanelProps) {
   return (
     <Box sx={styles.container}>
@@ -78,6 +81,7 @@ export default function StatisticsPanel({
           statisticsRecord={sr}
           rowIndex={rowIndex}
           removeStatistics={removeStatistics}
+          postMessage={postMessage}
         />
       ))}
     </Box>