+++ src/components/TimeSeriesPanel/TimeSeriesChartHeader.tsx
@@ -38,7 +38,6 @@ import FitScreenIcon from "@mui/icons-material/FitScreen";
 import IsoIcon from "@mui/icons-material/Iso";
 import ScatterPlotIcon from "@mui/icons-material/ScatterPlot";
 import ShowChartIcon from "@mui/icons-material/ShowChart";
-import CameraAltIcon from "@mui/icons-material/CameraAlt";
 
 import i18n from "@/i18n";
 import {
@@ -50,9 +49,9 @@ import { WithLocale } from "@/util/lang";
 import { makeStyles } from "@/util/styles";
 import { TimeSeriesChartType } from "@/states/controlState";
 import { MessageType } from "@/states/messageLogState";
+import SnapshotButton from "@/components/SnapshotButton";
 import TimeSeriesAddButton from "./TimeSeriesAddButton";
 import ValueRangeEditor from "./ValueRangeEditor";
-import { ExportOptions, exportElement } from "@/util/export";
 
 type ValueRange = [number, number];
 const SHOW_DEV_VALUE = "stddev";
@@ -86,7 +85,6 @@ const styles = makeStyles({
     paddingRight: theme.spacing(1),
   }),
 });
-const EXPORT_WIDTH = 2000; // in pixels
 
 interface TimeSeriesChartHeaderProps extends WithLocale {
   timeSeriesGroup: TimeSeriesGroup;
@@ -162,23 +160,6 @@ export default function TimeSeriesChartHeader({
       values.length === 1 ? (values[0] as TimeSeriesChartType) : chartType,
     );
     setStdevBars(showStdDevNew);
-  };
-
-  const handleExportSuccess = () => {
-    postMessage("success", i18n.get("Snapshot copied to clipboard"));
-  };
-
-  const handleExportError = (error: unknown) => {
-    const message = "Error copying snapshot to clipboard";
-    console.error(message + ":", error);
-    postMessage("error", i18n.get(message));
-  };
-
-  const exportOptions: ExportOptions = {
-    format: "png",
-    width: EXPORT_WIDTH,
-    handleSuccess: handleExportSuccess,
-    handleError: handleExportError,
   };
 
   return (
@@ -274,19 +255,7 @@ export default function TimeSeriesChartHeader({
           </Tooltip>
         </ToggleButtonGroup>
 
-        <Tooltip arrow title={i18n.get("Copy snapshot of chart to clipboard")}>
-          <IconButton
-            key={"exportButton"}
-            sx={styles.actionButton}
-            onClick={() =>
-              chartElement.current &&
-              exportElement(chartElement.current, exportOptions)
-            }
-            size="small"
-          >
-            <CameraAltIcon fontSize={"inherit"}></CameraAltIcon>
-          </IconButton>
-        </Tooltip>
+        <SnapshotButton elementRef={chartElement} postMessage={postMessage} />
 
         <TimeSeriesAddButton
           sx={styles.actionButton}